Fixed a bug with incorrect spawners
* Fixed a bug where cat spawners would be ocelot spawners.
