+++ pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>de.dustplanet</groupId>
     <artifactId>SilkSpawners</artifactId>
-    <version>6.1.3-TI-BETA</version>
+    <version>6.1.4</version>
     <packaging>jar</packaging>
 
     <name>SilkSpawners</name>
