+++ src/main/java/de/dustplanet/silkspawners/util/Common.java
@@ -1,6 +1,5 @@
 package de.dustplanet.silkspawners.util;
 
-
 import java.lang.reflect.Field;
 
 import org.bukkit.Bukkit;
@@ -8,14 +7,10 @@ import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandMap;
 import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
-import org.bukkit.scheduler.BukkitRunnable;
 
 import lombok.Getter;
 import lombok.Setter;
-import net.md_5.bungee.api.ChatMessageType;
-import net.md_5.bungee.api.chat.TextComponent;
 
 /**
  * @author kangarko
@@ -25,19 +20,6 @@ public final class Common {
     @Setter
     @Getter
     private static JavaPlugin instance;
-
-    public static void sendTitle(Player pl, String title, String subtitle) {
-        pl.sendTitle(colorize(title), colorize(subtitle), 20, 3 * 20, 10);
-    }
-
-
-    /*FIXME: public static void sendBar(Player pl, String title) {
-        try {
-            pl.spigot().sendMessage(ChatMessageType.ACTION_BAR, new TextComponent(colorize(title))); //barAPI?
-        } catch (final Throwable t) {
-            tell(pl, title);
-        }
-    }*/
 
     public static void log(String... messages) {
         for (final String message : messages)
@@ -72,13 +54,5 @@ public final class Common {
         } catch(final Exception e) {
             e.printStackTrace();
         }
-    }
-
-    public static <T> T getOrDefault(T nullable, T def) {
-        return nullable != null ? nullable : def;
-    }
-
-    public static void runLater(int delay, Runnable task) {
-        Bukkit.getScheduler().runTaskLater(instance, task, delay);
     }
 }
